Create UsersService with method for calculating user daily calories intake

+++ Data/StayFit.Data.Models/ApplicationUser.cs
@@ -50,7 +50,7 @@ namespace StayFit.Data.Models
 
         public ActivityLevel ActivityLevel { get; set; }
 
-        public decimal DailyCalories { get; set; }
+        public double DailyCalories { get; set; }
 
         public virtual ICollection<IdentityUserRole<string>> Roles { get; set; }
